# Replace `batched_convolution` by PyTensor native implementation