--- a/environment.yml
+++ b/environment.yml
@@ -16,7 +16,8 @@ dependencies:
 - preliz
 - pyprojroot
 # NOTE: Keep minimum pymc version in sync with ci.yml `OLDEST_PYMC_VERSION`
-- pymc>=5.21.1
+- pymc>=5.22.0
+- pytensor>=2.30.3
 - pymc-extras>=0.2.1
 - blackjax>=1.2.4
 - scikit-learn>=1.1.1