This will depend on the not yet released https://github.com/pymc-devs/pytensor/pull/1318

That will also require a dependency bump on PyMC before we can use it here.

The PyTensor impl gives us batching out of the box, without a graph compile and runtime penalty. I got these runtime differences before and after my changes: 

```python
import numpy as np
import pytensor.tensor as pt
import pytensor
from pymc_marketing.mmm.transformers import batched_convolution

for kernel_size in (8, 80, 800):
    for data_size in (12, 120, 1200):
        data = pt.vector("x", shape=(data_size,))
        kernel = pt.vector("kernel", shape=(kernel_size,))

        out = batched_convolution(data, kernel, mode="After")
        fn = pytensor.function([data, kernel], out, trust_input=True)

        rng = np.random.default_rng((kernel_size, data_size))
        data_test = rng.normal(size=data.type.shape)
        kernel_test = rng.normal(size=kernel.type.shape)

        print(f"{kernel_size=}, {data_size=}")
        %timeit fn(data_test, kernel_test)

# Before changes:
# kernel_size=8, data_size=12
# 35.3 μs ± 4.67 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# kernel_size=8, data_size=120
# 29.7 μs ± 432 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# kernel_size=8, data_size=1200
# 30.1 μs ± 391 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

# kernel_size=80, data_size=12
# 398 μs ± 1.82 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# kernel_size=80, data_size=120
# 402 μs ± 5.93 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# kernel_size=80, data_size=1200
# 401 μs ± 6.74 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# kernel_size=800, data_size=12
# 7.28 ms ± 45.7 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# kernel_size=800, data_size=120
# 7.97 ms ± 469 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# kernel_size=800, data_size=1200
# 7.5 ms ± 406 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# After changes:
# kernel_size=8, data_size=12
# 6.06 μs ± 61.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# kernel_size=8, data_size=120
# 6.32 μs ± 18.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# kernel_size=8, data_size=1200
# 9.18 μs ± 67.7 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

# kernel_size=80, data_size=12
# 6.67 μs ± 180 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# kernel_size=80, data_size=120
# 9.81 μs ± 60.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# kernel_size=80, data_size=1200
# 39.3 μs ± 1.15 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

# kernel_size=800, data_size=12
# 9.22 μs ± 21.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# kernel_size=800, data_size=120
# 24.6 μs ± 425 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# kernel_size=800, data_size=1200
# 179 μs ± 2.92 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```
Not seen here is also the widely different compile times for very large kernel_sizes, which in the old implementation would be a series of unrolled `set_subtensor`. Now the compile time is constant on the number of lags. It's also no longer needed to have statically known lags.

The gradient should provide a nice speedup as well, with room for further cleverness outside of pymc-marketing: https://github.com/pymc-devs/pytensor/issues/1320

In practice users won't be doing a crazy number of lags, so speedup on runtime is perhaps just those 3-4x I saw locally for `kernel_size=8`, and not some of the 1000x I got for other combinations. It may be a higher/ lower delta in other backends, feel free to check. 

The compile time and simpler graphs are what drove me to do these changes

<!-- readthedocs-preview pymc-marketing start -->
----
📚 Documentation preview 📚: https://pymc-marketing--1583.org.readthedocs.build/en/1583/

<!-- readthedocs-preview pymc-marketing end -->