@cluhmann 2025-03-26
Can we get some examples of where this helps us in the library?  And if you can provide a vignette of how to leverage this new convolution, that would be great too.  I see that you import `from pytensor.tensor.signal import convolve`, but then don't use it (directly). Is that where the magic is happening?