@ricardoV94 2025-03-27
> Can we get some examples of where this helps us in the library? And if you can provide a vignette of how to leverage this new convolution, that would be great too. I see that you import `from pytensor.tensor.signal import convolve`, but then don't use it (directly). Is that where the magic is happening?

There's nothing you need to do to leverage. It's used by the geometric adstock, so any MMM models with it (most) will benefit from the PyTensor implementation, which is better than the custom one we had (atleast 3-4x runtime as reported on the vignette) and much faster compile times.

The import issues you found is just unfinished code, it's supposed to use `convolve1d` not `convolve`