@cetagostini 2025-04-02
I see https://github.com/pymc-devs/pytensor/pull/1318 merge. Do what would be next to merge here? @ricardoV94

Do you need a hand?