@ricardoV94 2025-04-03
> I see [pymc-devs/pytensor#1318](https://github.com/pymc-devs/pytensor/pull/1318) merge. Do what would be next to merge here? @ricardoV94
> 
> Do you need a hand?

It needed a pymc release, did it now