@juanitorduz 2025-04-05
> Seems like there's another, perhaps unrelated failure, with the latest PyTensor

For reference:

```python
ERROR    pytensor.graph.rewriting.basic:basic.py:1757 Rewrite failure due to: local_shape_to_shape_i
ERROR    pytensor.graph.rewriting.basic:basic.py:1758 node: Shape(Blockwise{Subtensor{start:stop}, (i00),(),()->(o00)}.0)
ERROR    pytensor.graph.rewriting.basic:basic.py:1759 TRACEBACK:
ERROR    pytensor.graph.rewriting.basic:basic.py:1760 Traceback (most recent call last):
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/rewriting/basic.py", line 1968, in process_node
    fgraph.replace_all_validate_remove(  # type: ignore
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/features.py", line 634, in replace_all_validate_remove
    chk = fgraph.replace_all_validate(replacements, reason=reason, **kwargs)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/features.py", line 579, in replace_all_validate
    fgraph.replace(r, new_r, reason=reason, verbose=False, **kwargs)
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/fg.py", line 535, in replace
    self.change_node_input(
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/fg.py", line 456, in change_node_input
    self.import_var(new_var, reason=reason, import_missing=import_missing)
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/fg.py", line 323, in import_var
    self.import_node(var.owner, reason=reason, import_missing=import_missing)
  File "/opt/hostedtoolcache/Python/3.12.9/x64/lib/python3.12/site-packages/pytensor/graph/fg.py", line 388, in import_node
    raise MissingInputError(error_msg, variable=var)
pytensor.graph.utils.MissingInputError: Input 0 (<Scalar(int64, shape=())>) of the graph (indices start from 0), used to compute Add(<Scalar(int64, shape=())>, Sub.0), was not provided and not given a value. Use the PyTensor flag exception_verbosity='high', for more information on this error.
```

Any ideas why this could be?