@ricardoV94 2025-04-09
Found the issue, will be fixed by https://github.com/pymc-devs/pytensor/pull/1353