@juanitorduz 2025-04-10
> We should lower pin the last minor version of PyTensor separately from the pymc pin, because pymc doesn't pin it

sure! sounds good!